3341: Temporarily add headless IdP
This adds Headless IdP to the new frontend local config, which should
allow the acceptance tests to pass.

Currently lots are failing because you can't progress beyond the "select
documents" page, as the set of IdPs returned by the API from the old
frontend, and the set of IdPs the new frontend knows about, are
disjoint.